Move related resources view to its own dir
