+++ contentcuration/contentcuration/frontend/channelEdit/views/related-resources/RelatedResourcesView/related-resources-preview.png
